backend/requirements.txt
psycopg2-binary==2.9
pydantic==1.9.1
